+++ src/lightning/pytorch/callbacks/__init__.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 from lightning.pytorch.callbacks.batch_size_finder import BatchSizeFinder
 from lightning.pytorch.callbacks.callback import Callback
-from lightning.pytorch.callbacks.channels_last import ChannelsLast
+from lightning.pytorch.callbacks.memory_format import MemoryFormat
 from lightning.pytorch.callbacks.checkpoint import Checkpoint
 from lightning.pytorch.callbacks.device_stats_monitor import DeviceStatsMonitor
 from lightning.pytorch.callbacks.early_stopping import EarlyStopping
@@ -38,7 +38,6 @@ __all__ = [
     "BasePredictionWriter",
     "BatchSizeFinder",
     "Callback",
-    "ChannelsLast",
     "Checkpoint",
     "DeviceStatsMonitor",
     "EarlyStopping",
@@ -46,6 +45,7 @@ __all__ = [
     "LambdaCallback",
     "LearningRateFinder",
     "LearningRateMonitor",
+    "MemoryFormat",
     "ModelCheckpoint",
     "ModelPruning",
     "ModelSummary",